Lab 3, first try

+++ src/main/java/by/triumgroup/recourse/controller/impl/AbstractCrudController.java
@@ -1,38 +1,79 @@
 package by.triumgroup.recourse.controller.impl;
 
+import by.triumgroup.recourse.configuration.security.Auth;
+import by.triumgroup.recourse.configuration.security.UserAuthDetails;
 import by.triumgroup.recourse.controller.CrudController;
+import by.triumgroup.recourse.controller.exception.AccessDeniedException;
 import by.triumgroup.recourse.controller.exception.BadRequestException;
 import by.triumgroup.recourse.controller.exception.NotFoundException;
 import by.triumgroup.recourse.entity.model.BaseEntity;
+import by.triumgroup.recourse.entity.model.User;
 import by.triumgroup.recourse.service.CrudService;
+import by.triumgroup.recourse.service.UserService;
 import org.slf4j.Logger;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import javax.validation.Valid;
 import java.util.Optional;
+import java.util.function.BiFunction;
 
 import static by.triumgroup.recourse.util.ServiceCallWrapper.wrapServiceCall;
 
-public abstract class AbstractCrudController<E extends BaseEntity<ID>, ID> implements CrudController<E,ID> {
+public abstract class AbstractCrudController<E extends BaseEntity<ID>, ID> implements CrudController<E, ID> {
 
+    private final UserService userService;
     private final Logger logger;
     private final CrudService<E, ID> crudService;
 
-    AbstractCrudController(CrudService<E, ID> crudService, Logger logger) {
+    AbstractCrudController(CrudService<E, ID> crudService, UserService userService, Logger logger) {
+        this.userService = userService;
         this.logger = logger;
         this.crudService = crudService;
     }
 
-    public E getById(@PathVariable("id") ID id) {
-        return wrapServiceCall(logger, () -> {
-            Optional<E> callResult = crudService.findById(id);
-            return callResult.orElseThrow(NotFoundException::new);
-        });
+    protected abstract boolean hasAuthorityToEdit(E entity, UserAuthDetails authDetails);
+
+    protected boolean hasAuthorityToRead(E entity, UserAuthDetails authDetails) {
+        return true;
+    }
+
+    protected void checkAuthority(E entity, UserAuthDetails authDetails, BiFunction<E, UserAuthDetails, Boolean> authorityChecker) {
+        refreshAuthDetails(authDetails);
+        if (!authDetails.isAdmin() && !authorityChecker.apply(entity, authDetails)) {
+            throw new AccessDeniedException();
+        }
+    }
+
+    protected void refreshAuthDetails(UserAuthDetails authDetails) {
+        Optional<User> user = userService.findById(authDetails.getId());
+        if (user.isPresent()) {
+            authDetails.setRole(user.get().getRole());
+        } else {
+            throw new AccessDeniedException();
+        }
     }
 
     @Override
-    public <S extends E> S create(@Valid @RequestBody S entity) {
+    public E getById(@PathVariable("id") ID id, @Auth UserAuthDetails userAuthDetails) {
+        return wrapServiceCall(logger, () -> {
+            Optional<E> callResult = crudService.findById(id);
+            if (callResult.isPresent()) {
+                checkAuthority(callResult.get(), userAuthDetails, this::hasAuthorityToRead);
+                return callResult.get();
+            } else {
+                throw new NotFoundException();
+            }
+        });
+    }
+
+    public Iterable<E> getAll(@Auth UserAuthDetails authDetails) {
+        return wrapServiceCall(logger, crudService::findAll);
+    }
+
+    @Override
+    public <S extends E> S create(@Valid @RequestBody S entity, @Auth UserAuthDetails authDetails) {
+        checkAuthority(entity, authDetails, this::hasAuthorityToEdit);
         return wrapServiceCall(logger, () -> {
             Optional<S> callResult = crudService.add(entity);
             return callResult.orElseThrow(BadRequestException::new);
@@ -40,15 +81,17 @@ public abstract class AbstractCrudController<E extends BaseEntity<ID>, ID> imple
     }
 
     @Override
-    public <S extends E> S update(@Valid @RequestBody S entity, @PathVariable("id") ID id) {
+    public E update(@Valid @RequestBody E entity, @PathVariable("id") ID id, @Auth UserAuthDetails authDetails) {
+        checkAuthority(entity, authDetails, this::hasAuthorityToEdit);
         return wrapServiceCall(logger, () -> {
-            Optional<S> callResult = crudService.update(entity, id);
+            Optional<E> callResult = crudService.update(entity, id);
             return callResult.orElseThrow(NotFoundException::new);
         });
     }
 
     @Override
-    public void delete(@PathVariable("id") ID id) {
+    public void delete(@PathVariable("id") ID id, @Auth UserAuthDetails authDetails) {
+        checkAuthority(getById(id, authDetails), authDetails, this::hasAuthorityToEdit);
         wrapServiceCall(logger, () -> crudService.delete(id).orElseThrow(NotFoundException::new));
     }
 }
